requirements.txt
requests==2.22.0
flask==1.1.1
moment==0.8.2
python-telegram-bot==12.4.2